feat: huge improvements on perf
