+++ bin/ddrive.js
@@ -8,6 +8,7 @@ const {
     TIMEOUT = 60000, // Optional
     CHUNK_SIZE = 7864320, // 7.5MB
     SECRET, // Optional
+    CHUNK_PROCESS_CONCURRENCY = 3, // Upload this many chunks at once into discord
 } = process.env
 
 const startApp = async () => {
@@ -16,6 +17,7 @@ const startApp = async () => {
         webhooks: WEBHOOKS.split(','),
         chunkSize: CHUNK_SIZE,
         secret: SECRET,
+        chunkProcessConcurrency: CHUNK_PROCESS_CONCURRENCY,
         restOpts: {
             timeout: TIMEOUT,
         },
